+++ src/__tests__/rendererIntegrationTests/testAppClient.js
@@ -16,6 +16,7 @@ const spectron = require('spectron');
 //   testName: will be displayed in the window
 //   userKey: clients will be initialized with this user
 //   streaming: true to enable streaming
+//   useNetModule: true to use Electron's net module
 
 function testAppClient(fakeLD, baseOptions) {
   let app;
@@ -52,6 +53,7 @@ function testAppClient(fakeLD, baseOptions) {
     args.push('-u', options.userKey);
     args.push('-e', envs.join(','));
     options.streaming && args.push('--streaming');
+    options.useNetModule && args.push('--useNetModule');
 
     app = new spectron.Application({
       path: electron,