+++ src/electronPlatform.js
@@ -4,13 +4,20 @@ const { EventSource } = require('launchdarkly-eventsource');
 const newHttpRequest = require('./httpRequest');
 const packageJson = require('../package.json');
 
-function makeElectronPlatform(options) {
+function makeElectronPlatform(options, logger) {
   const tlsParams = filterTlsParams(options && options.tlsParams);
+  const useNetModule = options && options.useNetModule;
+
+  if (useNetModule && Object.keys(tlsParams).length > 0 && logger && typeof logger.warn === 'function') {
+    logger.warn(
+      "The useNetModule option is enabled, so tlsParams will be ignored. Electron's net module delegates TLS " +
+        'configuration to Chromium. Disable useNetModule to use custom TLS parameters.'
+    );
+  }
 
   const ret = {};
 
-  ret.httpRequest = (method, url, headers, body) =>
-    newHttpRequest(method, url, headers, body, tlsParams, options && options.useNetModule);
+  ret.httpRequest = (method, url, headers, body) => newHttpRequest(method, url, headers, body, tlsParams, useNetModule);
 
   ret.httpAllowsPost = () => true;