+++ src/index.js
@@ -11,10 +11,14 @@ const packageJson = require('../package.json');
 // initializeRenderer).
 function initializeInMain(env, user, options = {}) {
   // Pass our platform object to the common code to create the Electron version of the client
-  const platform = electronPlatform(options);
-  const extraOptionDefs = {};
+  const logger = options.logger || createDefaultLogger();
+  const platform = electronPlatform(options, logger);
+  const extraOptionDefs = {
+    tlsParams: { type: 'object' },
+    useNetModule: { default: false },
+  };
   if (!options.logger) {
-    extraOptionDefs.logger = { default: createDefaultLogger() };
+    extraOptionDefs.logger = { default: logger };
   }
   const clientVars = common.initialize(env, user, options, platform, extraOptionDefs);
   const client = clientVars.client;